+++ src/Http/Controllers/CsrfCookieController.php
@@ -2,6 +2,8 @@
 
 namespace Laravel\Sanctum\Http\Controllers;
 
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
 class CsrfCookieController
@@ -9,10 +11,15 @@ class CsrfCookieController
     /**
      * Return an empty response simply to trigger the storage of the CSRF cookie in the browser.
      *
+     * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function show()
+    public function show(Request $request)
     {
+        if ($request->expectsJson()) {
+            return new JsonResponse(null, 204);
+        }
+
         return new Response('', 204);
     }
 }