+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Release Notes
 
-## [Unreleased](https://github.com/laravel/airlock/compare/v2.8.2...master)
+## [Unreleased](https://github.com/laravel/sanctum/compare/v2.9.0...master)
+
+
+## [v2.9.0 (2021-01-26)](https://github.com/laravel/sanctum/compare/v2.8.2...v2.9.0)
+
+### Added
+- Add multiple guard support for SPA auth ([#246](https://github.com/laravel/sanctum/pull/246), [f5695ae](https://github.com/laravel/sanctum/commit/f5695aecc547138c76bc66aaede73ba549dabdc5))
+
+### Fixed
+- Return json response when the request expects a json ([#247](https://github.com/laravel/sanctum/pull/247))
 
 ### Added
 - Expiration Dates ([#252](https://github.com/laravel/sanctum/pull/252))